+++ app/models/login/dfe_login.rb
@@ -18,7 +18,7 @@ module Login
     end
 
     def logout_url(routable)
-      routable.supply_teachers_rm6238_gateway_url
+      routable.supply_teachers_rm6376_gateway_url
     end
 
     def permit?(framework)